@@ -22,12 +22,13 @@
     "eslint-plugin-prettier": "^5.5.1",
     "nostr-tools": "^2.15.0",
     "qrcode.react": "^4.2.0",
-    "react": "^19.1.0",
+    "react": "^19.2.3",
     "react-dnd": "^16.0.1",
     "react-dnd-html5-backend": "^16.0.1",
-    "react-dom": "^19.1.0",
+    "react-dom": "^19.2.3",
     "react-intl": "^7.1.11",
     "react-markdown": "^10.1.0",
+    "react-router": "^7.11.0",
     "react-select": "^5.10.2",
     "remark-gfm": "^4.0.1",
     "uuid": "^11.1.0",
@@ -35,8 +36,8 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.30.1",
-    "@types/react": "^19.1.8",
-    "@types/react-dom": "^19.1.6",
+    "@types/react": "^19.2.7",
+    "@types/react-dom": "^19.2.3",
     "@vitejs/plugin-react": "^4.6.0",
     "eslint": "^9.30.1",
     "eslint-plugin-react-hooks": "^5.2.0",
