@@ -1,12 +1,14 @@
-import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import "./index.css";
 import App from "./App";
+// @ts-expect-error ignore servie worker for now
 import * as serviceWorker from "./serviceWorker";
 
-setTimeout(() => {
-  ReactDOM.render(<App />, document.getElementById("root"));
-});
+const domNode = document.getElementById("root");
+if (domNode) {
+  const root = createRoot(domNode);
+  root.render(<App />);
+}
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
