@@ -8,6 +8,8 @@ import { IntlProvider } from "react-intl";
 import { flattenMessages } from "./common/utils";
 import dictionary from "./common/dictionary";
 import LoginModal from "./components/LoginModal";
+import { BrowserRouter } from "react-router";
+import { Routing } from "./components/Routing";
 
 let _locale =
   (navigator.languages && navigator.languages[0]) ||
@@ -15,7 +17,7 @@ let _locale =
   "en-US";
 _locale = ~Object.keys(dictionary).indexOf(_locale) ? _locale : "en-US";
 
-function App() {
+function Application() {
   const {
     user,
     isInitialized,
@@ -99,4 +101,10 @@ function App() {
   );
 }
 
-export default App;
+export default function App() {
+  return (
+    <BrowserRouter>
+      <Routing indexNode={<Application />} />
+    </BrowserRouter>
+  );
+}
